fix the topic list
